login COntroller class added
