+++ routes/loginRoutes/loginRoute.js
@@ -2,6 +2,7 @@ const express = require('express');
 const orm_login = require('../../ORM/ORM_login');
 const orm_emp_info = require('../../ORM/ORM_employee_info');
 const checkCookie = require('../../authentication/cookie_checker');
+const loginController = require('../../controllers/logincontroller'); // Import the loginController function
 require('dotenv').config();
 
 
@@ -11,9 +12,11 @@ const router = express.Router();
 router.post('/api/v0/login', async (req, res) => { //post->get
     // console.log("Backend: requested-> /api/v0/login");
     // req_json = req.body;
+    console.log("loginRoute");
     console.log(req);
     const pseudo_view = await loginController(req);
     if(pseudo_view) {
+        console.log('User found!');
         console.log(pseudo_view);
         res.json(pseudo_view);
     }
@@ -28,47 +31,7 @@ router.post('/api/v0/login', async (req, res) => { //post->get
 });
 
 
-async function loginController(req) {
-    // req_json.username = 'doc_oc'
-    // req_json.password = 'MTIz'
-    const req_json = req.body;
-    console.log("loginController")
-    console.log(req.body, req.query);
-    const model_user_login_info = await orm_login.verifyLogin(req_json.username, req_json.password);
-    if(model_user_login_info) {
-        console.log('User found!');
-        console.log(model_user_login_info);
-        // set cookie
-        session = req.session;
-        // session.userid = model_user_login_info.username;
-        let model = null;
-        if(model_user_login_info.user_role.toUpperCase() == 'DOCTOR') {
-            model = await orm_emp_info.get_doctor_info(req_json.username);
-        }
-        else if(model_user_login_info.user_role.toUpperCase() == 'INTERN') {
-            model = await orm_emp_info.get_intern_info(req_json.username);
-        }
-        else if(model_user_login_info.user_role.toUpperCase() == 'RECEPTIONIST') {
-            model = await orm_emp_info.get_receptionist_info(req_json.username);
-        }
 
-        console.log(model);
-
-        const pseudo_view = {
-            username : req_json.username,
-            role : model_user_login_info.user_role,
-            info : model,
-        };
-        
-        console.log(pseudo_view);
-
-        return pseudo_view;
-    }
-    else {
-        console.log('NO user Found!');
-        return null;
-    }
-}
